+++ tiger-testenv-mgr/src/test/java/de/gematik/test/tiger/testenvmgr/TestEnvManagerConfigurationCheck.java
@@ -22,13 +22,18 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.TestInstance.Lifecycle;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.springframework.test.util.ReflectionTestUtils;
+import uk.org.webcompere.systemstubs.environment.EnvironmentVariables;
+import uk.org.webcompere.systemstubs.jupiter.SystemStub;
+import uk.org.webcompere.systemstubs.jupiter.SystemStubsExtension;
 
 @Slf4j
 @Getter
 @TestInstance(Lifecycle.PER_CLASS)
+@ExtendWith(SystemStubsExtension.class)
 class TestEnvManagerConfigurationCheck {
 
   @BeforeEach
@@ -73,31 +78,29 @@ class TestEnvManagerConfigurationCheck {
       skipEnvironmentSetup = true)
   void testCheckCfgPropertiesMissingParamMandatoryServerPortProp_NOK(TigerTestEnvMgr envMgr) {
     CfgServer srv = envMgr.getConfiguration().getServers().get("testTigerProxy");
-    assertThatThrownBy(
-            () -> envMgr.createServer("testTigerProxy", srv).assertThatConfigurationIsCorrect())
+    AbstractTigerServer server = envMgr.createServer("testTigerProxy", srv);
+    assertThatThrownBy(server::assertThatConfigurationIsCorrect)
         .isInstanceOf(TigerTestEnvException.class);
   }
 
   @Test
   void testCheckDoubleKey_NOK() {
-    assertThatThrownBy(
-            () ->
-                TigerGlobalConfiguration.initializeWithCliProperties(
-                    Map.of(
-                        "TIGER_TESTENV_CFGFILE",
-                        "src/test/resources/de/gematik/test/tiger/testenvmgr/testDoubleKey.yaml")))
+    Map<String, String> yamlMap =
+        Map.of(
+            "TIGER_TESTENV_CFGFILE",
+            "src/test/resources/de/gematik/test/tiger/testenvmgr/testDoubleKey.yaml");
+    assertThatThrownBy(() -> TigerGlobalConfiguration.initializeWithCliProperties(yamlMap))
         .isInstanceOf(TigerConfigurationException.class)
         .hasMessage("Duplicate keys in yaml file ('serverDouble')!");
   }
 
   @Test
   void testCheckDeprecatedKey_port_NOK() {
-    assertThatThrownBy(
-            () ->
-                TigerGlobalConfiguration.initializeWithCliProperties(
-                    Map.of(
-                        "TIGER_TESTENV_CFGFILE",
-                        "src/test/resources/de/gematik/test/tiger/testenvmgr/testDeprecatedKey.yaml")))
+    Map<String, String> yamlMap =
+        Map.of(
+            "TIGER_TESTENV_CFGFILE",
+            "src/test/resources/de/gematik/test/tiger/testenvmgr/testDeprecatedKey.yaml");
+    assertThatThrownBy(() -> TigerGlobalConfiguration.initializeWithCliProperties(yamlMap))
         .isInstanceOf(TigerConfigurationException.class)
         .hasMessageContaining(
             "The key ('port') in yaml file should not be used anymore, use 'proxyPort' instead!");
@@ -105,12 +108,11 @@ class TestEnvManagerConfigurationCheck {
 
   @Test
   void testCheckDeprecatedKey_tigerport_NOK() {
-    assertThatThrownBy(
-            () ->
-                TigerGlobalConfiguration.initializeWithCliProperties(
-                    Map.of(
-                        "TIGER_TESTENV_CFGFILE",
-                        "src/test/resources/de/gematik/test/tiger/testenvmgr/testDeprecatedKey.yaml")))
+    Map<String, String> yamlMap =
+        Map.of(
+            "TIGER_TESTENV_CFGFILE",
+            "src/test/resources/de/gematik/test/tiger/testenvmgr/testDeprecatedKey.yaml");
+    assertThatThrownBy(() -> TigerGlobalConfiguration.initializeWithCliProperties(yamlMap))
         .isInstanceOf(TigerConfigurationException.class)
         .hasMessageContaining(
             "The key ('port') in yaml file should not be used anymore, use 'proxyPort' instead!");
@@ -118,12 +120,11 @@ class TestEnvManagerConfigurationCheck {
 
   @Test
   void testCheckDeprecatedKey_serverPort_NOK() {
-    assertThatThrownBy(
-            () ->
-                TigerGlobalConfiguration.initializeWithCliProperties(
-                    Map.of(
-                        "TIGER_TESTENV_CFGFILE",
-                        "src/test/resources/de/gematik/test/tiger/testenvmgr/testDeprecatedKey.yaml")))
+    Map<String, String> yamlMap =
+        Map.of(
+            "TIGER_TESTENV_CFGFILE",
+            "src/test/resources/de/gematik/test/tiger/testenvmgr/testDeprecatedKey.yaml");
+    assertThatThrownBy(() -> TigerGlobalConfiguration.initializeWithCliProperties(yamlMap))
         .isInstanceOf(TigerConfigurationException.class)
         .hasMessageContaining(
             "The key ('serverPort') in yaml file should not be used anymore, use 'adminPort'"
@@ -132,12 +133,11 @@ class TestEnvManagerConfigurationCheck {
 
   @Test
   void testCheckDeprecatedKey_proxyCfg_NOK() {
-    assertThatThrownBy(
-            () ->
-                TigerGlobalConfiguration.initializeWithCliProperties(
-                    Map.of(
-                        "TIGER_TESTENV_CFGFILE",
-                        "src/test/resources/de/gematik/test/tiger/testenvmgr/testDeprecatedKey.yaml")))
+    Map<String, String> yamlMap =
+        Map.of(
+            "TIGER_TESTENV_CFGFILE",
+            "src/test/resources/de/gematik/test/tiger/testenvmgr/testDeprecatedKey.yaml");
+    assertThatThrownBy(() -> TigerGlobalConfiguration.initializeWithCliProperties(yamlMap))
         .isInstanceOf(TigerConfigurationException.class)
         .hasMessageContaining(
             "The key ('proxyCfg') in yaml file should not be used anymore, it is omitted!");
@@ -145,12 +145,11 @@ class TestEnvManagerConfigurationCheck {
 
   @Test
   void testCheckDeprecatedKey_healthcheck_NOK() {
-    assertThatThrownBy(
-            () ->
-                TigerGlobalConfiguration.initializeWithCliProperties(
-                    Map.of(
-                        "TIGER_TESTENV_CFGFILE",
-                        "src/test/resources/de/gematik/test/tiger/testenvmgr/testDeprecatedKey.yaml")))
+    Map<String, String> yamlMap =
+        Map.of(
+            "TIGER_TESTENV_CFGFILE",
+            "src/test/resources/de/gematik/test/tiger/testenvmgr/testDeprecatedKey.yaml");
+    assertThatThrownBy(() -> TigerGlobalConfiguration.initializeWithCliProperties(yamlMap))
         .isInstanceOf(TigerConfigurationException.class)
         .hasMessageContaining(
             "The key ('tiger.servers.*.externalJarOptions.healthcheck') in yaml file should not be"
@@ -159,12 +158,11 @@ class TestEnvManagerConfigurationCheck {
 
   @Test
   void testCheckDeprecatedKey_healthcheckurl_NOK() {
-    assertThatThrownBy(
-            () ->
-                TigerGlobalConfiguration.initializeWithCliProperties(
-                    Map.of(
-                        "TIGER_TESTENV_CFGFILE",
-                        "src/test/resources/de/gematik/test/tiger/testenvmgr/testDeprecatedKey.yaml")))
+    Map<String, String> yamlMap =
+        Map.of(
+            "TIGER_TESTENV_CFGFILE",
+            "src/test/resources/de/gematik/test/tiger/testenvmgr/testDeprecatedKey.yaml");
+    assertThatThrownBy(() -> TigerGlobalConfiguration.initializeWithCliProperties(yamlMap))
         .isInstanceOf(TigerConfigurationException.class)
         .hasMessageContaining(
             "The key ('tiger.servers.*.externalJarOptions.healthcheckurl') in yaml file should not"
@@ -202,6 +200,48 @@ class TestEnvManagerConfigurationCheck {
         .isEqualTo(TigerGlobalConfiguration.readIntegerOptional("free.port.4").get());
     assertThat(tigerServer2.getConfiguration().getTigerProxyCfg().getProxiedServerProtocol())
         .isEqualTo("ftp");
+  }
+
+  @SystemStub
+  private final EnvironmentVariables serverVariable =
+      new EnvironmentVariables("ENV_VAR", "httpbin");
+
+  @Test
+  @TigerTest(
+      tigerYaml =
+          """
+          servers:
+            httpbinName:
+              startupTimeoutSec: 1
+              type: ${ENV_VAR}
+          """)
+  void testEnvironmentVariables(TigerTestEnvMgr envMgr) {
+    final AbstractTigerServer httpbin = envMgr.getServers().get("httpbinName");
+    assertThat(httpbin.getServerTypeToken()).isEqualTo("httpbin");
+  }
+
+  @SystemStub
+  private final EnvironmentVariables timVariables =
+      new EnvironmentVariables("TIM_KEYSTORE", "src\\test\\resources\\egk_aut_keystore.jks")
+          .and("TIM_KEYSTORE_PW", "gematik");
+
+  @Test
+  @TigerTest(
+      tigerYaml =
+          """
+          tigerProxy:
+            tls:
+              forwardMutualTlsIdentity: "${TIM_KEYSTORE};${TIM_KEYSTORE_PW};00"
+                  """)
+  void testProxyEnvironmentVariables(TigerTestEnvMgr envMgr) {
+    assertThat(
+            envMgr
+                .getLocalTigerProxyOrFail()
+                .getTigerProxyConfiguration()
+                .getTls()
+                .getForwardMutualTlsIdentity()
+                .getFileLoadingInformation())
+        .isEqualTo("src\\test\\resources\\egk_aut_keystore.jks;gematik;00");
   }
 
   // -----------------------------------------------------------------------------------------------------------------